trying to fix speed resets when the fan turns itself on and notifies homekit about the change
I think the issue arises when the device is turned off and homekit gets the speedvalue from getSpeed(). homebridge returns 0 as the device is off. when the device turns on, it notifies homebridge which then notifies homekit. However, homekit is not notified about the speed change thus still thinking it is at 0% speed.