implemented a fix for resetting values on device turn on
issue: when homekit calles getValue when a device is off, 0 is returned but cachedValue is not updated. when the device is turned back on, homekit resets its value to 100 or leaves it at 0, not turning on
implemented fix: do not reset the value to 0 at all when getValue is called. return the cached value or, if undefined, 100

fix still needs testing